// src/websites/komorii/src/types/api.ts
export interface SeriesSearchResult {
    seriesId: string;
    title: string;
    seedCount?: number;
    chapterCount?: number;
    lastUploadedAt?: string; // ISO Date
    source: number;
    externalMangaId: string;
    year?: number;
    latestChapterNumber?: number;
    latestChapterTitle?: string;
}